# docs/swagger.yaml
definitions:
  cluster.APIUser:
    properties:
      grants:
        additionalProperties:
          type: boolean
        type: object
      roles:
        additionalProperties:
          type: boolean
        type: object
      user:
        type: string
    type: object
  cluster.Agent:
    properties:
      cpuCores:
        type: integer
      cpuFreq:
        type: integer
      hostName:
        type: string
      id:
        type: string
      memBytes:
        type: integer
      memFreeBytes:
        type: integer
      osKernel:
        type: string
      osName:
        type: string
      status:
        type: string
      version:
        type: string
    type: object
...
  cluster.Cluster:
    properties:
      activePassiveStatus:
        type: string
      agents:
        items:
          $ref: '#/definitions/cluster.Agent'
        type: array
      apiUsers:
        additionalProperties:
          $ref: '#/definitions/cluster.APIUser'
        type: object
      backupList:
        $ref: '#/definitions/config.BackupMetaMap'
      backupStat:
        $ref: '#/definitions/repmanv3.BackupStat'
      blacklist:
        items:
          $ref: '#/definitions/regexp.Regexp'
        type: array
      canConnectVault:
        type: boolean
      canInitNodes:
        type: boolean
      cleanReplication:
        description: used in testing
        type: boolean
      config:
        $ref: '#/definitions/github_com_signal18_replication-manager_config.Config'
      configurator:
        $ref: '#/definitions/configurator.Configurator'
      dbServers:
        items:
          type: string
        type: array
      dbServersCrashes:
        description: This will be purged on all db node up
        items:
          $ref: '#/definitions/cluster.Crash'
        type: array
      diffVariables:
        items:
          $ref: '#/definitions/cluster.VariableDiff'
        type: array
      diskType:
        additionalProperties:
          type: string
        type: object
      failoverCounter:
        type: integer
      failoverHistory:
        description: This will be used for PITR
        items:
          $ref: '#/definitions/cluster.Crash'
        type: array
      failoverLastTime:
        type: integer
      fsType:
        additionalProperties:
          type: boolean
        type: object
      haveDBTLSCert:
        type: boolean
      haveDBTLSOldCert:
        type: boolean
      inBinlogBackup:
        type: boolean
      inLogicalBackup:
        type: boolean
      inPhysicalBackup:
        type: boolean
      inResticBackup:
        type: boolean
      inRollingRestart:
        type: boolean
      isAlertDisable:
        type: boolean
      isAllDbUp:
        type: boolean
      isCapturing:
        type: boolean
      isClusterDown:
        type: boolean
      isDown:
        type: boolean
      isExportPush:
        type: boolean
      isFailable:
        type: boolean
      isFailedArbitrator:
        type: boolean
      isGettingSlowLog:
        type: boolean
      isGitPull:
        type: boolean
      isGitPush:
        type: boolean
      isLostMajority:
        type: boolean
      isMasterDown:
        type: boolean
      isNeedDatabasesConfigChange:
        type: boolean
      isNeedDatabasesReprov:
        type: boolean
      isNeedDatabasesRestart:
        type: boolean
      isNeedDatabasesRollingReprov:
        type: boolean
      isNeedDatabasesRollingRestart:
        type: boolean
      isNeedProxiesConfigChange:
        type: boolean
      isNeedProxiesRestart:
        type: boolean
      isNeedProxyRestart:
        type: boolean
      isNotMonitoring:
        type: boolean
      isPostgres:
        type: boolean
      isProvision:
        type: boolean
      isSplitBrain:
        type: boolean
      isValidBackup:
        type: boolean
      jobResults:
        $ref: '#/definitions/config.TasksMap'
      lastDelayStatPrint:
        type: string
      log:
        $ref: '#/definitions/s18log.HttpLog'
      logTask:
        $ref: '#/definitions/s18log.HttpLog'
      monitorSpin:
        type: string
      monitorType:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      proxyServers:
        items:
          type: string
        type: array
      slaHistory:
        items:
          $ref: '#/definitions/state.Sla'
        type: array
      slavesConnected:
        type: integer
      slavesOldestMasterFile:
        $ref: '#/definitions/cluster.SlavesOldestMasterFile'
      sqlErrorLog:
        $ref: '#/definitions/s18log.HttpLog'
      sqlGeneralLog:
        $ref: '#/definitions/s18log.HttpLog'
      sstAvailablePorts:
        additionalProperties:
          type: string
        type: object
      stateMachine:
        allOf:
        - $ref: '#/definitions/state.StateMachine'
        description: |-
          dbUser                        string                      `json:"-"`
          oldDbUser string `json:"-"`
          dbPass                        string                      `json:"-"`
          oldDbPass string `json:"-"`
          rplUser                   string                      `json:"-"`
          rplPass                   string                      `json:"-"`
          proxysqlUser              string                      `json:"-"`
          proxysqlPass              string                      `json:"-"`
      tenant:
        type: string
      topology:
        type: string
      topologyType:
        additionalProperties:
          type: string
        type: object
      uptime:
        type: string
      uptimeFailable:
        type: string
      uptimeSemisync:
        type: string
      useBlacklist:
        type: boolean
      useWhitelist:
        type: boolean
      versionsMap:
        $ref: '#/definitions/config.VersionsMap'
      vmType:
        additionalProperties:
          type: boolean
        type: object
      waitingFailover:
        type: integer
      waitingRejoin:
        type: integer
      waitingSwitchover:
        type: integer
      whitelist:
        items:
          $ref: '#/definitions/regexp.Regexp'
        type: array
      workLoad:
        $ref: '#/definitions/config.WorkLoad'
      workingDir:
        type: string
    type: object
...
  cluster.Crash:
    properties:
      electedMasterURL:
        type: string
      failoverIOGtid:
        items:
          $ref: '#/definitions/gtid.Gtid'
        type: array
      failoverMasterLogFile:
        type: string
      failoverMasterLogPos:
        type: string
      failoverSemiSyncSlaveStatus:
        type: boolean
      newMasterLogFile:
        type: string
      newMasterLogPos:
        type: string
      switchover:
        type: boolean
      unixTimestamp:
        type: integer
      url:
        type: string
    type: object
...
  cluster.UserForm:
    properties:
      grants:
        type: string
      password:
        type: string
      roles:
        type: string
      username:
        type: string
    type: object
...
  config.BackupMetaMap:
    type: object
...
  config.TasksMap:
    type: object
  config.VersionsMap:
    type: object
...
  regexp.Regexp:
    type: object
  repmanv3.BackupStat:
    properties:
      total_blob_count:
        type: integer
      total_file_count:
        type: integer
      total_size:
        type: integer
    type: object
...
  server.Heartbeat:
    properties:
      cluster:
        type: string
      failed:
        type: integer
      hosts:
        type: integer
      id:
        type: integer
      master:
        type: string
      secret:
        type: string
      status:
        type: string
      uuid:
        type: string
    type: object
...
  server.token:
    properties:
      token:
        type: string
    type: object
  server.userCredentials:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /api/clusters:
    get:
      description: Fetches the list of clusters that the user has access to based
        on ACL.
      produces:
      - application/json
      responses:
        "200":
          description: List of clusters
          schema:
            items:
              $ref: '#/definitions/cluster.Cluster'
            type: array
        "401":
          description: Unauthenticated resource
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Fetch clusters
      tags:
      - Cluster
...
  /api/clusters/{clusterName}/actions/master-physical-backup:
    post:
      consumes:
      - application/json
      description: This endpoint triggers a physical backup for the master of the
        specified cluster.
      parameters:
      - description: Cluster Name
        in: path
        name: clusterName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully triggered physical backup
          schema:
            type: string
        "400":
          description: No cluster found
          schema:
            type: string
        "403":
          description: No valid ACL
          schema:
            type: string
      summary: Perform a physical backup for the master of a specific cluster
      tags:
      - ClusterBackup
...
  /api/clusters/{clusterName}/servers/{serverName}/{serverPort}/backup:
    get:
      description: Initiates a physical backup on a specified server port within a
        cluster.
      parameters:
      - description: Cluster Name
        in: path
        name: clusterName
        required: true
        type: string
      - description: Server Name
        in: path
        name: serverName
        required: true
        type: string
      - description: Server Port
        in: path
        name: serverPort
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Backup initiated successfully
          schema:
            type: string
        "403":
          description: No valid ACL
          schema:
            type: string
        "500":
          description: Cluster Not Found" or "Server Not Found
          schema:
            type: string
      summary: Perform a physical backup on a server port
      tags:
      - DatabaseBackup
...
  /api/clusters/{clusterName}/servers/{serverName}/{serverPort}/is-master:
    get:
      description: Checks if a specified server port within a cluster is a master.
      parameters:
      - description: Cluster Name
        in: path
        name: clusterName
        required: true
        type: string
      - description: Server Name
        in: path
        name: serverName
        required: true
        type: string
      - description: Server Port
        in: path
        name: serverPort
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: 200 -Valid Master!
          schema:
            type: string
        "500":
          description: No cluster
          schema:
            type: string
        "503":
          description: 503 -Not a Valid Master!
          schema:
            type: string
      summary: Check if a server port is a master
      tags:
      - Database
  /api/clusters/{clusterName}/servers/{serverName}/{serverPort}/is-slave:
    get:
      description: Checks if a specified server port within a cluster is a slave.
      parameters:
      - description: Cluster Name
        in: path
        name: clusterName
        required: true
        type: string
      - description: Server Name
        in: path
        name: serverName
        required: true
        type: string
      - description: Server Port
        in: path
        name: serverPort
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: 200 -Valid Slave!
          schema:
            type: string
        "500":
          description: No cluster
          schema:
            type: string
        "503":
          description: 503 -Not a Valid Slave!
          schema:
            type: string
      summary: Check if a server port is a slave
      tags:
      - Database
...
  /api/clusters/{clusterName}/servers/{serverName}/is-master:
    get:
      description: Checks if a specified server within a cluster is a master.
      parameters:
      - description: Cluster Name
        in: path
        name: clusterName
        required: true
        type: string
      - description: Server Name
        in: path
        name: serverName
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: 200 -Valid Master!
          schema:
            type: string
        "500":
          description: No cluster
          schema:
            type: string
        "503":
          description: 503 -Not a Valid Master!
          schema:
            type: string
      summary: Check if a server is a master
      tags:
      - Database
  /api/clusters/{clusterName}/servers/{serverName}/is-slave:
    get:
      description: Checks if a specified server within a cluster is a slave.
      parameters:
      - description: Cluster Name
        in: path
        name: clusterName
        required: true
        type: string
      - description: Server Name
        in: path
        name: serverName
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: 200 -Valid Slave!
          schema:
            type: string
        "500":
          description: No cluster
          schema:
            type: string
        "503":
          description: 503 -Not a Valid Slave!
          schema:
            type: string
      summary: Check if a server is a slave
      tags:
      - Database
  /api/clusters/{clusterName}/servers/{serverName}/master-status:
    get:
      description: Retrieves the master status of a specified server within a cluster.
      parameters:
      - description: Cluster Name
        in: path
        name: clusterName
        required: true
        type: string
      - description: Server Name
        in: path
        name: serverName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Master status retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "403":
          description: No valid ACL
          schema:
            type: string
        "500":
          description: Cluster Not Found" or "Server Not Found" or "Encoding error
          schema:
            type: string
      summary: Get master status of a server
      tags:
      - Database
...
  /api/clusters/{clusterName}/servers/{serverName}/status:
    get:
      description: Retrieves the status of a specified server within a cluster.
      parameters:
      - description: Cluster Name
        in: path
        name: clusterName
        required: true
        type: string
      - description: Server Name
        in: path
        name: serverName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Status retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "403":
          description: No valid ACL
          schema:
            type: string
        "500":
          description: Cluster Not Found" or "Server Not Found" or "Encoding error
          schema:
            type: string
      summary: Get status of a server
      tags:
      - Database
...
  /api/clusters/{clusterName}/status:
    get:
      description: This endpoint retrieves the status of a specified cluster and returns
        it in JSON format.
      parameters:
      - description: Cluster Name
        in: path
        name: clusterName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Cluster status
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: No cluster found
          schema:
            type: string
      summary: Retrieve status of a cluster
      tags:
      - Cluster
  /api/clusters/{clusterName}/subscribe:
    post:
      consumes:
      - application/json
      description: This endpoint allows a user to subscribe to a specified cluster.
      parameters:
      - description: Cluster Name
        in: path
        name: clusterName
        required: true
        type: string
      - description: User Form
        in: body
        name: userform
        required: true
        schema:
          $ref: '#/definitions/cluster.UserForm'
      produces:
      - application/json
      responses:
        "200":
          description: Email sent to admin!
          schema:
            type: string
        "400":
          description: Error in request
          schema:
            type: string
        "401":
          description: 'Error logging in to gitlab: Token credentials is not valid'
          schema:
            type: string
        "403":
          description: Error parsing JWT" / "Current user is not logged in via Gitlab!
          schema:
            type: string
        "409":
          description: User already subscribed on peer cluster!" / "Another user already
            subscribed on peer cluster!
          schema:
            type: string
        "500":
          description: No valid cluster" / "Peer does not have cloud18 setup!" / "Error
            sending email
          schema:
            type: string
      summary: Subscribe a user to a cluster
      tags:
      - Cloud18
...
  /api/clusters/{clusterName}/topology/servers:
    get:
      description: This endpoint retrieves the servers for the specified cluster.
      parameters:
      - description: Cluster Name
        in: path
        name: clusterName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of servers
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Retrieve servers for a specific cluster
      tags:
      - ClusterTopology
...
  /api/clusters/for-sale:
    get:
      description: This endpoint returns a list of peer clusters that are available
        for sale, excluding those that are booked by the requesting user.
      parameters:
      - description: JWT token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of peer clusters available for sale
          schema:
            items:
              $ref: '#/definitions/config.PeerCluster'
            type: array
        "401":
          description: Unauthenticated resource
          schema:
            type: string
        "500":
          description: Failed to get token claims or Error Marshal
          schema:
            type: string
      summary: Retrieve peer clusters for sale
      tags:
      - Cloud18
  /api/clusters/peers:
    get:
      description: This endpoint retrieves the peer clusters that a user has access
        to.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of peer clusters
          schema:
            items:
              $ref: '#/definitions/config.PeerCluster'
            type: array
        "401":
          description: Unauthenticated resource
          schema:
            type: string
        "500":
          description: Failed to get token claims or Error Marshal
          schema:
            type: string
      summary: Retrieve peer clusters for a user
      tags:
      - Cloud18
...
  /api/configs/grafana:
    get:
      description: Returns a list of Grafana files from the specified directory.
      produces:
      - application/json
      responses:
        "200":
          description: List of Grafana files
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List Grafana files
      tags:
      - Public
  /api/heartbeat:
    get:
      consumes:
      - application/json
      description: Returns the heartbeat status of the replication manager.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.Heartbeat'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Monitor Heartbeat
      tags:
      - Public
  /api/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token upon successful login.
      parameters:
      - description: User credentials
        in: body
        name: userCredentials
        required: true
        schema:
          $ref: '#/definitions/server.userCredentials'
      produces:
      - application/json
      responses:
        "200":
          description: JWT token
          schema:
            $ref: '#/definitions/server.token'
        "401":
          description: Invalid credentials
          schema:
            type: string
        "403":
          description: Error in request
          schema:
            type: string
        "429":
          description: Too many requests
          schema:
            type: string
        "500":
          description: Error signing token
          schema:
            type: string
      summary: User login
      tags:
      - Auth
  /api/monitor:
    get:
      consumes:
      - application/json
      description: This endpoint processes the replication manager requests, validates
        cluster ACLs, and returns the cluster list in JSON format.
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with replication manager details
          schema:
            $ref: '#/definitions/server.ReplicationManager'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Handles replication manager requests
      tags:
      - Public
  /api/prometheus:
    get:
      description: Fetches Prometheus metrics for all servers in all clusters managed
        by the replication manager.
      produces:
      - text/plain
      responses:
        "200":
          description: Prometheus metrics
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Fetch Prometheus metrics
      tags:
      - Public
  /api/repocomp/current:
    get:
      description: Reads the current repository component from the specified directory
        and returns its content.
      produces:
      - text/plain
      responses:
        "200":
          description: Current repository component content
          schema:
            type: string
        "404":
          description: 404 Something went wrong - Not Found
          schema:
            type: string
      summary: Retrieve current repository component
      tags:
      - Public
...
swagger: "2.0"